+++ tests/integration/listassign/test.php
@@ -15,7 +15,7 @@ foo( $mixed );
 
 list( $safe, $unsafe ) = $mixed;
 
-// This is safe, but is not reported as such because array elements share taintedness
+
 echo $safe;
 echo $unsafe;