+++ tests/integration/weirdsyntax/test9.php
@@ -1,0 +1,3 @@
+<?php
+
+list (0) = array(3); // No crash